requirements.txt
cryptography==41.0.7
SQLAlchemy @ git+https://github.com/sqlalchemy/sqlalchemy.git@main
python-dotenv==1.0.1
bcrypt==4.1.2
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
requests==2.31.0
python-dateutil==2.8.2
Flask-Migrate==4.0.7
Waitress==3.0.0